+++ c/dsls/Makefile.no-autoconfig
@@ -123,9 +123,11 @@ lemon: lemon.c \
   lemon_symbol.c \
   lemon_plink.h \
   lemon_plink.c \
+  lemon_parse.h \
+  lemon_parse.c \
   lemon_report.h \
   lemon_report.c
-	$(DSLCC) -o lemon lemon.c lemon_assert.c lemon_error.c lemon_string.c lemon_memory.c lemon_option.c lemon_action.c lemon_symbol.c lemon_plink.c lemon_report.c
+	$(DSLCC) -o lemon lemon.c lemon_assert.c lemon_error.c lemon_string.c lemon_memory.c lemon_option.c lemon_action.c lemon_symbol.c lemon_plink.c lemon_parse.c lemon_report.c
 
 # ----------------------------------------------------------------
 clean:
