+++ c/dsls/lemon_parse.h
@@ -1,0 +1,8 @@
+#ifndef LEMON_PARSE_H
+#define LEMON_PARSE_H
+
+#include "lemon_structs.h"
+
+int Parse(struct lemon *lemp, int nDefine, char** azDefine);
+
+#endif // LEMON_PARSE_H
